feat: added methods recursively materializing OtapBatch ids and parent ids
